# backend/src/main/resources/expert-management-openapi.yaml
openapi: 3.0.3
info:
  title: Expert Management System API
  version: 1.0.0
  description: API do zarządzania kompetencjami pracowników.
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: [ ]
paths:
...
  /users/{id}:
    get:
      tags: [ Users ]
      summary: Pobierz użytkownika po ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Szczegóły użytkownika
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [ Users ]
      summary: Aktualizuj użytkownika
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Użytkownik zaktualizowany
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags: [ Users ]
      summary: Usuń użytkownika
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Użytkownik usunięty
...
  /domains/{id}:
    put:
      tags: [ Domains ]
      summary: Aktualizuj domenę
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainRequest'
      responses:
        '200':
          description: Domena zaktualizowana
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
    delete:
      tags: [ Domains ]
      summary: Usuń domenę
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Domena usunięta
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - active
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
...
    User:
      type: object
      required:
        - firstName
        - lastName
        - email
        - active
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        active:
          type: boolean
    Domain:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
...
    DomainRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string